Add (header)brower ,piblciIp,sms/ussd auth with featureIp/userFeatureIp

+++ src/main/java/com/gl/ceir/config/model/app/User.java
@@ -19,13 +19,11 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
- 
-    private long id;
+    private int id;
 
     String username;
-
     String password;
-
     int parentId;
+
 
 }
